--- a/examples/mnist_swwae.py
+++ b/examples/mnist_swwae.py
@@ -65,7 +65,8 @@ def convresblock(x, nfeats=8, ksize=3, nskipped=2, elu=True):
         if elu:
             y = ELU()(y)
         else:
-            y = BatchNormalization(mode=0, axis=1)(y)
+            # This example assume 'channels_first' data format.
+            y = BatchNormalization(axis=1)(y)
             y = Activation('relu')(y)
         y = Convolution2D(nfeats, 1, 1, border_mode='same')(y)
     return merge([y0, y], mode='sum')