@fchollet 2017-03-12
Like, the API for the BN layer has changed...