@fchollet 2017-03-12
I've update this example in `keras-2` to follow the new API. Check it out.