@nouiz 2017-03-13
If you have a recent Theano version and cudnn, you should not see as much
slow down from using BN.

Le lun. 13 mars 2017 04:15, antonmbk <notifications@github.com> a écrit :

> @fchollet <https://github.com/fchollet> I follow now, I goofed and
> somehow got the branch in the wrong way, will grab it the right way
> tomorrow and fix ASAP.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/fchollet/keras/pull/5722#issuecomment-286040524>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AALC-xExFWzPabk0fZFZ7oCxIVC_2KoKks5rlPsSgaJpZM4MaZCZ>
> .
>