@antonmbk 2017-03-14
@nouiz I  believe I am using recent Theano + Cudnn, but will double check.  I am going to submit another PR that branches correctly (shame on me) and then maybe before these changes get incorporated you can run the examples to let me know what timing was on your machine.  Either way I like leaving the option of ELU or BN+RELU as they both have advantages and disadvantages.