@antonmbk 2017-03-14
Resubmitted with correct 2.0 branch