@isaacgerg 2017-08-11
@antonmbk I believe you can now do the unpooling with tensorflow.  Have you looked in to this?  I would like to add it with your help.