@antonmbk 2017-08-14
@isaacgerg  Check out https://github.com/fchollet/keras/issues/4694